ath79: add WIP support for D-Link DAP-1320 / DAP-1120
The DAP-1320 is an N300 repeater that comes in a tiny wall plug case.

Specifications:

 * AR9341 / QCA9533, 8 MiB Flash, 64 MiB RAM
 * Since this device has no ethernet port, wifi is enabled on boot
   using the default credentials as printed on the label
 * DAP-1120 A1 is the US black case variant of DAP-1320 Bx

Installation:

 * Web UI: http://192.168.0.50 (or different address obtained via DHCP)
   Also reachable via dlinkap.local and dlinkapXXXX.local,
   XXXX being the last octets of the MAC address, as seen in the ESSID
   There is no password set by default
 * Recovery Web UI: Keep reset button pressed during power-on
   until LED starts flashing red, upgrade via http://192.168.0.50

This is work-in-progress and still needs some polishing regarding changes
to device support (e.g. dts syntax) introduced within the last two years.

Also failsafe mode is currently not accessible,
if you lock yourself out of the device, just reflash using the Recovery.

Signed-off-by: Sebastian Schaper <openwrt@sebastianschaper.net>